refactor: add widget when add data table
