+++ apps/web/src/common/modules/widgets/_store/widget-generate-store.ts
@@ -246,6 +246,7 @@ export const useWidgetGenerateStore = defineStore('widget-generate', () => {
         /* Step 2 */
         reset: () => {
             state.widget = undefined;
+            state.dataTables = [];
             state.showOverlay = false;
             state.overlayStep = 1;
             state.selectedDataTableId = undefined;
